Create netflix_challenge.py
This file provides a solution for a movie recommendation challenge. Using data analysis, it offers personalized movie recommendations based on user preferences. It utilizes a modified version of collaborative filtering algorithm to analyze user ratings and similarities, providing accurate suggestions for movies that users are likely to enjoy. It uses both user-user and item-item collaborative filtering.